requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pandas==2.2.3
requests==2.32.3
openai==1.50.0
pydantic==2.9.2
python-dotenv==1.0.0
python-docx==1.1.0
python-multipart==0.0.9
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==10.2.0

